requirements.txt
# Core LangChain dependencies
langchain
langchain-openai
langchain-text-splitters
langchain-community
langgraph

# Database dependencies
sqlalchemy
langgraph-checkpoint-sqlite

# Vector store and embeddings
langchain_chroma
chromadb
openai

# Data processing
pandas
numpy

# Environment and configuration
python-dotenv

# Jupyter notebook dependencies
jupyter
ipykernel

# Evaluation and testing
pytest
pytest-asyncio

# Additional utilities
requests
python-dateutil
PyPDFLoader
